refactor: move dashboard plugin to `node-app`
