+++ imports/node-app/plugins/dashboard/i18n/ar.json
