+++ imports/node-app/plugins/dashboard/i18n/bg.json
