+++ imports/node-app/plugins/dashboard/i18n/es.json
