+++ imports/node-app/plugins/dashboard/i18n/fr.json
