+++ imports/node-app/plugins/dashboard/i18n/nb.json
