+++ imports/node-app/plugins/dashboard/i18n/sl.json
