+++ imports/node-app/plugins/dashboard/i18n/zh.json
