+++ imports/node-app/plugins/dashboard/index.js
@@ -1,4 +1,4 @@
-import i18n from "./i18n";
+import i18n from "./i18n/index.js";
 
 /**
  * @summary Import and call this function to add this plugin to your API.
